+++ packages/components/src/FieldBase/_FieldBase.tsx
@@ -44,6 +44,7 @@ const _FieldBase = <T extends ElementType>(
     description,
     errorMessage,
     className,
+    stateProps,
     ...rest
   } = props;
   const classNames = useClassNames({
@@ -57,6 +58,7 @@ const _FieldBase = <T extends ElementType>(
       ref={ref}
       className={cn('group/field', twWidth[width], classNames, className)}
       data-required={props.isRequired ? true : undefined}
+      data-error={props.isInvalid ? true : undefined}
       {...rest}
     >
       {label ? (