+++ themes/theme-core/src/tokens.ts
@@ -47,7 +47,7 @@ export const colors = {
     disabled: '#cccccc',
 
     // Status
-    success: '#008000',
+    success: '#008800',
     error: '#f33',
     info: '#008',
     warning: '#d80',
@@ -126,6 +126,7 @@ export const colors = {
     hover: '#8d8d8d',
 
     // Status
+    success: '#008800',
     error: '#f33',
     info: '#008',
     warning: '#d80',